pipenv: upgrade Pipenv
The current version is two years old and is causing some of our
packages installation to fail with messages such as:

  Installing dependencies from Pipfile.lock (63c834)...  error in
        <somepackage> setup command: 'extras_require' must be
        a dictionary whose values are strings or lists of strings
        containing valid project/version requirement specifiers.

With the relevant section in said package:

    extras_require={
        "Flask": ["Flask"],
        "dev": [
            "pytest",
            "pytest-flakes",
            "pytest-helpers-namespace",
        ],
    },

There is no clear violation of the expected format, however the
package is installed from a Git source. That's all I've gathered at
this point and I can't find a clear fix but this issue:

https://github.com/pypa/pipenv/issues/4309

Upgrading the pipenv version does fix the problem, so upgrade to the
latest possible version and roll with it.